Change to a simple line limit of 99 characters.
It was decided to ditch the soft 80/hard 100 in favour of just 100.

Now although people speak of figures like 80, 79 is actually the number
they mean, because terminals will tend to need an extra, blank line if
you write 80-character-wide lines.

This is reflected in PEP 8:

> Limit all lines to a maximum of 79 characters.

> …

> Some teams strongly prefer a longer line length. For code maintained exclusively or primarily by a team that can reach agreement on this issue, it is okay to increase the nominal line length from 80 to 100 characters (effectively increasing the maximum length to 99 characters), provided that comments and docstrings are still wrapped at 72 characters.

Therefore, having decided to go with 100, we’re actually going with 99.