Rework the license guideline
This is more accurate to my understanding of Rust licensing, and is
representative of the approach we have taken.

I removed most analysis of the license and left it an explanation of
how to pick a license that is compatible with Rust.

This flips the recommendation about MIT vs. Apache. MIT is the less
restrictive license and is more compatible with Rust than Apache. We
don't accept Apache only software in rust-lang projects, but we do
accept MIT-only.