Remove section on Organization
This section was vague and probably unactionable. It basically just exhorts the
reader to be "sensible", and gives a couple specific examples of where Serde
does somthing that is claimed to be sensible. We can reintroduce this section if
we come up with a way to give firmer advice here.