Avoid using "into" when explaining `to_*`
The description "Conversions prefixed `to_`, on the other hand, typically stay at the same level of abstraction but do some work to change one representation into another." risks overlapping the ideas of `to_` and `into_` to memory. Instead using "from... to" helps reduce that overlap by using to with `to_`.